### 2022-06-23 #4501 @editorialbot
[PRE REVIEW]: ALPPACA - A tooL for Prokaryotic Phylogeny And Clustering Analysis
**Submitting author:** @hkaspersen (<a href="http://orcid.org/0000-0002-9559-1303">Håkon Kaspersen</a>)
**Repository:** https://github.com/NorwegianVeterinaryInstitute/ALPPACA
**Branch with paper.md** (empty if default branch): joss_paper
**Version:** 1.0.4
