### 2022-06-24 #4501 @Kevin-Mattheus-Moerman
@hkaspersen thanks for this submission. I have added the `query-scope` label just now. This means I am labeling this work for a scope review by our editorial board. This is normal for works of this size and we will review if this work is in scope in terms of research application and functionality offered and if it complies with our ["substantial scholarly effort" criteria](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort). This scope review should take about 2 weeks to complete.
