### 2022-07-29 #4501 @Kevin-Mattheus-Moerman
@hkaspersen I have just removed the `query-scope` label since we've determined this may be in scope for JOSS. We will now move on to try to assign a handling editor for this work. Note however there is a backlog at the moment and most editors and handling many other submissions at the moment. Hence I'll waitlist this submission for now and we will assign an editor once one becomes available.
