### 2022-08-02 #4501 @csoneson
👋🏻 @coughls, @mberacochea - would you be interested in reviewing this submission for JOSS?

> ALPPACA - A tooL for Prokaryotic Phylogeny And Clustering Analysis
