### 2022-08-03 #4501 @csoneson
@mberacochea Brilliant, thanks! The way it works is that once we have secured two reviewers, I will open a review issue here on GitHub, and in there you will generate your review checklist (see [here](https://joss.readthedocs.io/en/latest/review_checklist.html) for what it looks like). You will then start going through the submission, checking off the items that you feel are satisfied and raising any concerns related to the others. It is often easier to fill the checklist and post issues "as you go", rather than waiting until you have reviewed the whole package. Usually we ask reviewers to provide a first review within two weeks of opening the review issue; however, I am of course aware that it's holiday season. It's helpful to have an estimate of how things are going though, and an estimated time when the review might be submitted.
So to summarize, feel free to start looking at the submission already, but please wait until we have recruited a second reviewer and the review issue is opened to post your comments (it's easier that way, only one place to look). More information about reviewing for JOSS is [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html), and of course feel free to ask if you have any questions!
