### 2022-08-13 #4501 @csoneson
👋🏻 @rcannood, @hseabolt - would one of you be interested in reviewing this submission for JOSS?

> ALPPACA - A tooL for Prokaryotic Phylogeny And Clustering Analysis
