### 2022-08-14 #4677 @editorialbot
[REVIEW]: ALPPACA - A tooL for Prokaryotic Phylogeny And Clustering Analysis
**Submitting author:** @hkaspersen (<a href="http://orcid.org/0000-0002-9559-1303">Håkon Kaspersen</a>)
**Repository:** https://github.com/NorwegianVeterinaryInstitute/ALPPACA
**Branch with paper.md** (empty if default branch): joss_paper
**Version:** v2.0.1
