### 2022-09-09 #4677 @hseabolt
@csoneson I have competed my initial review and have updated the review checklist in this thread as much as I am able to verify.  I do have additional specific comments for the authors -- should those comments and my recommendation be posted here directly or should those be sent to you to collect and share with the authors?

Thanks!
