### 2022-09-10 #4677 @hseabolt
@csoneson Thank you!  My review is below.

## General comments for the authors of ALPPACA, a Nextflow pipeline designed for prokaryotic phylogenetic and clustering analysis.

- The Nextflow code looks overall well designed and implemented -- why did the authors choose to include Shell wrapper scripts around Nextflow rather than just calling Nextflow and specifying the CLI arguments directly?
  This Shell wrapper layer is not typical of Nextflow pipelines and in my view, since these do not contain clear GetOpts or help messages for the user to understand them, they reduce the usability of this pipeline.

- This pipeline shares alot of overlap in functionality with the nf-core Bactmap pipeline (https://nf-co.re/bactmap), and it is not clear in the manuscript text how this pipeline's feature set or functionality compares to Bactmap or other similar pipelines.  It should be included in the text which new or different features set this pipeline apart from other ones currently available.

- Unfortunately, while I am able to successfully install the pipeline and obtain the example test data, I cannot verify successfully running the pipeline with any of the three tracks using either `test.sh` or `alppaca.sh` (additional details below).

- Specific comments I have are detailed below.

## Documentation

- The installation and how-to-run documentation for this pipeline are not mentioned either on its main Github page nor in the manuscript.  I did find the full documentation on Github as a Git Wiki, but please consider adding clear references to where this documentation can be found on the main Github page and/or in the manuscript text.

- Please include some additional detail on what solution(s) this pipeline is intended to offer in README on the main Github repo.  This could easily be adapted from the Statement of Need in the manuscript text, which is nicely written.

## Config files, profiles, dependencies, and container handling

- Please consider creating a `conf` folder in the main repo to hold additional custom configs users may wish to create (e.g. for other HPC schedulers, test profiles, etc).  Please additionally consider reorganizing the files in the main directory to more appropriate locations (e.g. `.sh` files should relocate to `bin`, `.config` files should be placed in `conf` with the exception of `nextflow.config`.
	Excellent examples of this style of repo organization can be found on any of the `nf-core` pipelines, e.g. viralrecon: https://github.com/nf-core/viralrecon

- I am not a fan of the user having to manually download container images and manually update the Nextflow config files, as I feel this defeats one of the key shareability aspects of using Nextflow.  DSL2 functionality enables container handling within modules, which would also have a secondary benefit of not needing to edit `nextflow.config` or either of the config files for the Shell wrappers.

- This pipeline is configured to only work with Singularity and Slurm, however Nextflow has native support for a number of container engines and compute environments.  While technically acceptable, I do feel this limitation detracts from the overall value of this work. I am not able to test the `cluster` profile since my HPC runs on Grid Engine.

- Please consider including a Conda `env.yml` in the repo that contains the Nextflow and Java versions required to successfully run this pipeline.   This will assist users who wish to run on HPC cluster environments but who may not have proper permissions or experience with installing software like Java in shared compute environments.

## Attempts to execute ALPPACA

- The example data was able to be downloaded successfully.

- The `test.sh` script does not work 'out of the box' due to the Nextflow binary being called `nextflow_21.10.6` on Line 24 of that script.  This should be updated to just `nextflow` since most users will not have their executables named consistently with the authors' system.
	Even with this minor change, the `test.sh` program does not execute successfully since the `test.config` file contains hard-coded filepaths on the authors' system.  The user is not instructed to update these in the relevant documentation sections.
	I tested all three tracks in the same manner following instructions given here: https://github.com/NorwegianVeterinaryInstitute/ALPPACA/wiki/1.-Installation.

- The same execution issues arise when trying to run the `alppaca.sh` script, since again there are hard-coded paths in the `main.config` file.  I have not made substantial effort to update these filepaths since there are several of them in each config file and this is a key barrier to the usability of ALPPACA -- in my judgement, most users will not go to the effort of manually updating these configs for every run and will simply look for a different solution.

- For Shell scripts in the main directory and in `bin`, please update the shebang lines to `#!/usr/bin/env bash`, which is more flexible for different environments your users may have.

## Recommendation

My recommendation is for the authors to make **minor to moderate revisions** (mainly to the codebase) and resubmit this manuscript and code repository.
