### 2022-09-17 #4677 @mberacochea
# My review of ALPPACA.

ALPPACA is a well written Nextflow pipeline, the code is clean and easy to follow. The paper is very well written and clearly states
the problem the authors are trying to solve.

## Documentation

The documentation of the software is stored in the GitHub Wiki section, but there are no links to it on either the manuscript or the repo README. Less experienced users will struggle to find it, please add a link to the README.

It would also be beneficial to use Read The Docs, Github Pages or a similar service to maintain the documentation and the output examples (which are a great resource).

Overall, the documentation is well written and covers the installation, functionality, and execution instructions of the software.

## Source code

As previously stated, the Nextflow code is well-designed and implemented. But, there are some hardcoded bits which require of some manual tweaking before running the pipeline.

Some of the hardcoded values:
- Nextflow binary (https://github.com/NorwegianVeterinaryInstitute/ALPPACA/blob/master/test.sh#L24)
- Slurm Cluster configuration values (https://github.com/NorwegianVeterinaryInstitute/ALPPACA/blob/master/nextflow.config#L81)

Please consider removing the paths and other hardcoded values.

Bash wrappers are used to run the pipeline (`alppaca.sh` and `test.sh`), this is an unusual pattern on Nextflow pipelines. In my experience such aproach is detrimental to the user experience. Also, the wrappers don't handle the arguments (e.g. GetOps) or have a help message.

## Execution

The documentation makes emphasis on Slurm as the scheduler, but Nextflow supports a wide range of Executors (https://www.nextflow.io/docs/latest/executor.html#executors). As it stands the only requirements are Nextflow and Singularity. I think it would be beneficial to follow nf-core configuration practices (https://nf-co.re/usage/configuration#basic-configuration-profiles). Such approach allows users to customize the execution of the pipeline easily and following a well establish practice.

I also consider that following an improved configuration scheme will help manage the containers, at the moment users need to download them manually which is not very user-friendly.

## Test execution

Following the instructions on the documentation (after fixing the nextflow hardcoded binary in `test.sh`) it was possible to run the pipeline on my laptop. I do consider that supporting Docker or Podman will make it the pipeline more portable, albeit Singularity is a popular containers tech on HPC clusters.

Please, include the Nextflow, Singularity and Java versions required to run the pipeline.

## Recommendation

My recommendation is for the authors to make minor revisions and resubmit the application.
