### 2022-09-19 NorwegianVeterinaryInstitute/ALPPACA#27 @rcannood
Comments on JOSS submission
This is my review to your JOSS submission (openjournals/joss-reviews#4677). I apologise for being the typical reviewer number three, but I do provide these comments with the intention of improving the manuscript and codebase.

**Manuscript**: While the author's work could definitely be of value to the community, I believe the JOSS submission should be reworked to highlight the specific qualities of the work. More specifically, the manuscript focuses a lot on technical aspects, e.g. it is a Nextflow pipeline with tools wrapped in Singularity containers, it generates an HTML report, the "Tracks" section reads like a methodology section of the tools the pipeline consists of.
The interesting academic parts of the manuscript would be, "What scientific problem does the work solve? How does it compare to other approaches in the field? What are the unique attributes of this work? What do the results of this pipeline look like? What can a user do with these results?"

**Codebase**: In addition, the current codebase is set up to be only useful for the author's specific HPC infrastructure, whereas Nextflow is designed to be used on any kind of system. As far as I know, there are no automated tests to evaluate whether the software is performing as intended.

In the comments below, I hope I have provided the seeds for how the authors might start to tackle some of the comments. I'd also be happy to hear from @csoneson on whether some of these topics are not part of my reviewing duties :)

I did not verify the installation instructions or functionality as it requires an unreasonable amount of effort, including setting up / acquiring my own HPC infrastructure and refactoring the author's codebase to remove any mentions of the author's HPC infrastructure and replace them with my own.

If done correctly, Nextflow pipelines are generic and can be run on a variety of systems, such as HPC clusters, Kubernetes cloud clusters or simply a beefy computer. For example, I should not need to:

* Manually clone the git repository: If you parameterise your pipeline correctly, you can use `nextflow run NorwegianVeternaryInstitute/ALPPACA <your parameters>` to initialise an execution
* Need to manually pull the Singularity images: by specifying the `container` directive in your Nextflow processes, Nextflow will automatically fetch the correct Singularity images and store them in the `NXF_SINGULARITY_CACHEDIR` folder.
* Edit the pipeline input parameters specified in `main.config`; I should just be able to pass the parameters at the commandline.
* Edit the `nextflow.config` to remove the author's HPC-specific settings such as their account username.

In my opinion, the authors should refactor the installation instructions so it lists something along the lines of:

Step 1: Create a nextflow.config with your infrastructure setup. Here is a template you can start from.
Step 2: Run `nextflow run NorwegianVeternaryInstitute/ALPPACA -c nextflow.config <your input parameters>`.

- [x] #32
- **Functionality > Functionality**: Will verify when above comment gets solved.
- **Functionality > Performance**: Will verify when above comment gets solved.

* [x] #33
* [x] #34
* [x] #35
* [x] #36

* [x] #38
  The authors should provide a high level description of what problems they try to solve and how they did it.

- [x] #39

- [x] #40

- [x] #41
