### 2022-09-19 #4677 @rcannood
> @mberacochea, @rcannood - could you update us on the state of your reviews? Thanks!

I created an issue in the ALPPACA repository. @csoneson Feel free to chime in on whether some of these comments are not required for the manuscript's acceptance at JOSS.
