### 2022-09-19 #4677 @hkaspersen
Dear @mberacochea, @hseabolt, @rcannood, and @csoneson.
Thank you so much for taking the time for this review, your feedback is invaluable! I am relatively new to Nextflow, so this will be a learning process for me. The feedback regarding the Nextflow standards is especially useful - I will try to make the pipeline more in line with these standards.

If you have any useful resources to share regarding Nextflow standards, please do not hesitate to share! Thanks in advance!

@csoneson is there a time limit for addressing these changes?
