### 2022-09-19 NorwegianVeterinaryInstitute/ALPPACA#27 @hkaspersen
@rcannood Thank you so much for the valuable feedback! I will come back to you when I have necessary changes to the pipeline.
