### 2022-09-19 #4677 @csoneson
> @csoneson is there a time limit for addressing these changes?

No direct time limit - in order to respect the reviewers' time I'd just like to make sure that the submission is indeed "actively worked on" (i.e., not set aside and stalled for months without updates). It would be great to have a few updates along the way and perhaps an estimate of when you think there may be a new version for the reviewers to look at, but no stress - the main goal is to have a piece of software that you're happy with 🙂

> Also, should the changes made to the pipeline based on the comments above be registered as issues on the ALPPACA github page?

That's one option, that makes it easy to track - you can also summarize the changes made here afterwards (and maybe point to specific commits addressing them).

Feel free to ping me at any time if you have questions!
