### 2022-09-19 NorwegianVeterinaryInstitute/ALPPACA#28 @hkaspersen
Fix dependency handling
Dependencies (singularity images) are as of today pre-downloaded by the user. The pipeline should instead be configured to handle this automatically, as mentioned in the JOSS review thread (https://github.com/openjournals/joss-reviews/issues/4677).

Remove the `setup.sh` script and let Nextflow handle the images directly, both for singularity and docker.
