### 2022-09-19 NorwegianVeterinaryInstitute/ALPPACA#29 @hkaspersen
remove wrapper script for running the pipeline
To follow proper Nextflow running, remove the wrapper script and run the pipeline directly with `nextflow run` instead. This is better since the user can pass parameter arguments directly in the run instead of manually changing the values in the `main.config` file, as mentioned in the JOSS review (https://github.com/openjournals/joss-reviews/issues/4677).

Is it then still possible to output a version report and a run report in the output directory, that shows which version of the pipeline was run, and which parameters were used?
