### 2022-09-19 NorwegianVeterinaryInstitute/ALPPACA#30 @hkaspersen
restructure config files
The pipeline should be able to run without needing to manually change values in a config file, as mentioned in the JOSS review (https://github.com/openjournals/joss-reviews/issues/4677). The config files should hold the recommended, default values for different analyses, while the user should only have to change parameters that deviate from the default when running the pipeline run command. See `nf-core` pipelines for good examples.
