### 2022-09-21 db05be8
removed local image paths and movedthe address to images to module definitions, as of issue #28
