### 2022-09-25 4d1de22
restructured the config to be more in line with nf-core standards, as of issue #30
