### 2022-09-25 NorwegianVeterinaryInstitute/ALPPACA#29 @hkaspersen
For version reporting, it is possible to add the following to the end of each process script:
```
cat <<-END_VERSIONS > versions.yml
"${task.process}":
cutadapt: \$(cutadapt --version)
END_VERSIONS
```
(example taken from https://github.com/nf-core/viralrecon/blob/master/modules/local/cutadapt.nf)
