### 2022-10-04 1fb1b36
placed container definitions back to the modules, didn't work with conda def. As of issue #28
