### 2022-10-04 NorwegianVeterinaryInstitute/ALPPACA#31 @hkaspersen
add an env.yml
Add `env.yml` to define a conda environment to run the pipeline with, defining both Java and Nextflow, as mentioned in https://github.com/openjournals/joss-reviews/issues/4677#issuecomment-1242767318
