### 2022-10-04 794de17
added env.yml file for nextflow environment, as of issue #31
