+++ assets/env.yml
@@ -1,0 +1,8 @@
+name: nextflow
+channels:
+ - conda-forge
+ - bioconda
+ - defaults
+dependencies:
+ - java-jdk=7.0.91=1
+ - nextflow=21.10.6=h4a94de4_0
