### 2022-10-04 NorwegianVeterinaryInstitute/ALPPACA#30 @hkaspersen
Old, manual config file removed. Added a new example config file in `conf/base.config` that users may use as a template for their systems, as mentioned in https://github.com/NorwegianVeterinaryInstitute/ALPPACA/issues/27#issue-1377416044.
