### 2022-10-04 NorwegianVeterinaryInstitute/ALPPACA#35 @hkaspersen
Automated tests: There are no automated tests to verify the functionality of the software. I understand that the pipeline requires an HPC to run the full pipeline, but automated tests could be made possible by unit testing the individual modules and/or integration testing on a subset of the original data.
Part of ALPPACA review: https://github.com/openjournals/joss-reviews/issues/4677
