### 2022-10-04 NorwegianVeterinaryInstitute/ALPPACA#28 @hkaspersen
Need to be able to use conda with the R-reporting step, asked about this in https://github.com/nextflow-io/nextflow/discussions/3265#discussion-4444128.
