### 2022-10-05 NorwegianVeterinaryInstitute/ALPPACA#28 @hkaspersen
The conda environment file `*.yml` can be called directly like this:
```
process foo {
  conda '/some/path/my-env.yaml'

  '''
  your_command --here
  '''
}
```
