### 2022-10-05 c4bcb57
added conda env yml for R reporting process
