+++ assets/r_env.yml
@@ -1,4 +1,4 @@
-name: my-env
+name: r_env
 channels:
   - conda-forge
   - bioconda
