+++ bin/core_genome_report_dedup.Rmd
@@ -1,288 +1,0 @@
----
-title: "ALPPACA Core Genome Report"
-date: "`r Sys.Date()`"
-output:
-  html_document:
-    mathjax: null
-    code_folding: "hide"
-    toc: true
-    toc_float:
-      collapsed: false
-params:
-  parsnp_report:
-    value: x
-  seqkit_report:
-    value: x
-  snpdist_report:
-    value: x
-  phylo_tree:
-    value: x
-  phylo_data:
-    value: x
----
-
-# Version information
-ALPPACA version 1.0.0.
-
-## R Packages
-These are the R packages used to generate this report.
-```{r, message=FALSE, warning=FALSE}
-library(stringr)
-library(readr)
-library(dplyr)
-library(tidyr)
-library(rlang)
-library(ggtree)
-library(plotly)
-library(ggplot2)
-library(viridis)
-library(formattable)
-library(kableExtra)
-
-packages <- (.packages())
-
-clean_pkg_version <- function(x) {
-  y <- getNamespaceVersion(x)
-  
-  z <- data.frame("Package" = x,
-                  "Version" = y)
-  
-  row.names(z) <- NULL
-  
-  return(z)
-}
-
-lapply(packages, function(x) clean_pkg_version(x)) %>%
-  bind_rows() %>%
-  kbl() %>%
-  kable_styling(bootstrap_options = c("striped",
-                                      "hover",
-                                      "responsive"),
-                full_width = TRUE)
-```
-
-<br>
-
-## Tool versions
-These are the tools used to run the analysis within the core genome workflow of ALPPACA.
-```{r, message=FALSE, warning=FALSE}
-tool_data <- data.frame(
-  Tool = c("ParSNP",
-           "Seqkit",
-           "Gubbins",
-           "MaskRC",
-           "Snp-dists",
-           "Snp-sites",
-           "IQTree"),
-  Version = c("1.6.1",
-              "2.1.0",
-              "3.1.3",
-              "0.5",
-              "0.8.2",
-              "2.5.1",
-              "2.1.4")
-)
-
-tool_data %>%
-  kbl() %>%
-  kable_styling(bootstrap_options = c("striped",
-                                      "hover",
-                                      "responsive"),
-                full_width = TRUE)
-```
-
-<br>
-
-<br>
-
-# Tool information
-## ParSNP
-
-<center>*ParSNP average coverage report. The "Type" column refers to which of the samples were used as a reference in the analysis. The "Percent Coverage" column represents the % coverage of each genome that was used to generate the core genome alignment. The green bar is relative to 100%. The total coverage is the average across all samples. Samples with expected high similarity (e.g. outbreak or same ST) are expected to have a high percent coverage.*</center>
-
-```{r, message=FALSE, warning=FALSE}
-# Courtesy of Eve Zeyl Fiskebeck for the script!
-# Read PARSNP results file
-parsnp_data <- readLines(params$parsnp_report)
-
-# Extract the data needed
-parsnp_cov <- parsnp_data[str_detect(parsnp_data, "(Cluster coverage in sequence)|(Total coverage among all sequences)")]
-parsnp_samples <- parsnp_data[str_detect(parsnp_data, "(Sequence)|(q value)")]
-
-df_samples <- as.data.frame(cbind(parsnp_samples)) %>%
-  head(-1) %>%
-  separate("parsnp_samples", into = c("Sequence", "Sample"), sep = " :") %>%
-  mutate(Sample = basename(Sample))
-
-# Get number of samples in the data
-number_samples <- nrow(df_samples)
-
-# Create nice table
-as.data.frame(cbind(parsnp_cov)) %>%
[…]
