+++ bin/mapping_report_dedup.Rmd
@@ -1,262 +1,0 @@
----
-title: "ALPPACA Mapping Report"
-date: "`r Sys.Date()`"
-output:
-  html_document:
-    mathjax: null
-    code_folding: "hide"
-    toc: true
-    toc_float:
-      collapsed: false
-params:
-  snippy_report:
-    value: x
-  seqkit_report:
-    value: x
-  snpdist_report:
-    value: x
-  phylo_tree:
-    value: x
-  phylo_data:
-    value: x
----
-# Run information
-
-Date generated: `r Sys.Date()`
-
-## R Packages used
-These are the R packages used to generate this report.
-```{r, message=FALSE, warning=FALSE}
-library(stringr)
-library(readr)
-library(dplyr)
-library(tidyr)
-library(rlang)
-library(ggtree)
-library(plotly)
-library(ggplot2)
-library(viridis)
-library(formattable)
-library(kableExtra)
-
-packages <- (.packages())
-
-clean_pkg_version <- function(x) {
-  y <- getNamespaceVersion(x)
-  
-  z <- data.frame("Package" = x,
-                  "Version" = y)
-  
-  row.names(z) <- NULL
-  
-  return(z)
-}
-
-lapply(packages, function(x) clean_pkg_version(x)) %>%
-  bind_rows() %>%
-  kbl() %>%
-  kable_styling(bootstrap_options = c("striped",
-                                      "hover",
-                                      "responsive"),
-                full_width = TRUE)
-```
-
-<br>
-
-## Tool versions
-These are the tools used to run the analysis within the core genome workflow of ALPPACA.
-```{r, message=FALSE, warning=FALSE}
-tool_data <- data.frame(
-  Tool = c("Snippy",
-           "Seqkit",
-           "Gubbins",
-           "MaskRC",
-           "Snp-dists",
-           "Snp-sites",
-           "IQTree"),
-  Version = c("4.6.0",
-              "2.1.0",
-              "3.2",
-              "0.5",
-              "0.8.2",
-              "2.5.1",
-              "2.1.4")
-)
-
-tool_data %>%
-  kbl() %>%
-  kable_styling(bootstrap_options = c("striped",
-                                      "hover",
-                                      "responsive"),
-                full_width = TRUE)
-```
-
-<br>
-
-<br>
-
-# Tool information
-## Snippy
-```{r, message=FALSE, warning=FALSE, fig.width=10, fig.height=6, dpi=600}
-snippy_data <- read_delim(params$snippy_report,
-                          delim = "\t",
-                          show_col_types = FALSE)
-
-options(scipen = 999)
-
-alignment_length <- unique(snippy_data$LENGTH)
-
-snippy_data %>%
-  select(-LENGTH) %>%
-  pivot_longer(cols = -ID,
-               values_to = "value",
-               names_to = "names") %>%
-  ggplot(aes(ID, value, fill = names)) +
-  geom_col() +
-  labs(x = NULL,
-       y = "Nucleotides",
-       fill = NULL) +
-  theme_bw() +
-  theme(axis.text.x = element_text(angle = 90, hjust = 1, vjust = 0.3))
-
-```
-
-<center>***Figure 1**: Summary statistics of the alignment generated by Snippy from mapped reads. The total alignment length was `r alignment_length` nucleotides.*</center>
-
-<br>
-
-## Deduplication
-```{r}
-if (file.size(params$seqkit_report) == 0L) {show_table <- FALSE} else {show_table <- TRUE}
-```
-`r if (show_table == FALSE){"No duplicated sequences detected."} else {"<center>*List of duplicated sequences identified in the multiple alignment.*</center>"}`
-
[…]

[…]
