### 2022-10-06 NorwegianVeterinaryInstitute/ALPPACA#37 @hkaspersen
For Shell scripts in the main directory and in bin, please update the shebang lines to #!/usr/bin/env bash, which is more flexible for different environments your users may have.
As mentioned in https://github.com/openjournals/joss-reviews/issues/4677#issuecomment-1242767318
