### 2022-10-06 8c09fe3
added new shebang, as of issue #37
