+++ assets/r_env.yml
@@ -4,6 +4,7 @@ channels:
   - bioconda
   - defaults
 dependencies:
+  - r-base=4.0.5
   - bioconductor-ggtree=3.2.0
   - r-stringr=1.4.0
   - r-readr=2.1.2
