### 2022-10-07 30a2839
added version output for each module, as of issue #29
