### 2022-10-07 #4677 @hkaspersen
Dear @mberacochea, @hseabolt, and @rcannood.
I am currently working on implementing some sort of test for the pipeline, but I am having a difficult time in figuring out how to do it the best way. If you have any suggestions or examples, please let me know! I see that `nf-core` provide tests that run on github actions, but I am not sure if I will manage to implement this here, as I am not familiar with this functionality.
