### 2022-10-07 #4677 @csoneson
> @csoneson can I update the version of the pipeline based on the changes I have made here without it affecting the review?

Yes - the version in the first post above is the version at the time of submission. You can update during the review process (but it won't be updated in this thread), and at the time of acceptance we sync the two again.
