### 2022-10-07 #4677 @hseabolt
@hkaspersen you can include some small test data files in the `assets` dir in your repo, and create some `test.config` files in `conf` dir.  Check out the example I have here: https://github.com/CDCgov/mycosnp-nf to see the file structure and examples of the `test.config` file.  You will need to import this config into your `nextflow.config` in the base dir, examples of this can be seen in the MycoSNP `nextflow.config`, Line 95.

If you set this up correctly, you can activate the tests at runtime with `nextflow run NorwegianVeterinaryInstitute/ALPPACA -profile test [other args]` and it shouldn't require any Github actions.  Note that you can give a multiple profiles to `profile` as a csv, e.g. `-profile docker,test`.
