### 2022-10-09 153c04f
added test config, issue #35
