### 2022-10-11 bdc41d1
fixed the yml with correct packages and versions, as of issue#28
