+++ assets/r_env.yml
@@ -1,18 +1,218 @@
-name: r_env
+name: alppaca_report
 channels:
   - conda-forge
   - bioconda
   - defaults
 dependencies:
-  - r-base=4.0.5
-  - bioconductor-ggtree=3.2.0
-  - r-stringr=1.4.0
-  - r-readr=2.1.2
-  - r-dplyr=1.0.8
-  - r-tidyr=1.2.0
-  - r-plotly=4.10.0
-  - r-ggplot2=3.3.5
-  - r-viridis=0.6.2
-  - r-formattable=0.2.1
-  - r-kableExtra=1.3.4
-  - r-patchwork=1.1.1
+  - _libgcc_mutex=0.1=conda_forge
+  - _openmp_mutex=4.5=2_gnu
+  - _r-mutex=1.0.1=anacondar_1
+  - binutils_impl_linux-64=2.36.1=h193b22a_2
+  - binutils_linux-64=2.36=hf3e587d_10
+  - bioconductor-ggtree=3.2.0=r41hdfd78af_0
+  - bioconductor-treeio=1.18.0=r41hdfd78af_0
+  - bwidget=1.9.14=ha770c72_1
+  - bzip2=1.0.8=h7f98852_4
+  - c-ares=1.18.1=h7f98852_0
+  - ca-certificates=2022.9.24=ha878542_0
+  - cairo=1.16.0=ha12eb4b_1010
+  - curl=7.85.0=h2283fc2_0
+  - expat=2.4.9=h27087fc_0
+  - font-ttf-dejavu-sans-mono=2.37=hab24e00_0
+  - font-ttf-inconsolata=3.000=h77eed37_0
+  - font-ttf-source-code-pro=2.038=h77eed37_0
+  - font-ttf-ubuntu=0.83=hab24e00_0
+  - fontconfig=2.14.0=hc2a2eb6_1
+  - fonts-conda-ecosystem=1=0
+  - fonts-conda-forge=1=0
+  - freetype=2.12.1=hca18f0e_0
+  - fribidi=1.0.10=h36c2ea0_0
+  - gcc_impl_linux-64=9.5.0=h6c5bc03_16
+  - gcc_linux-64=9.5.0=h4258300_10
+  - gettext=0.19.8.1=h27087fc_1009
+  - gfortran_impl_linux-64=9.5.0=h3c9b8b6_16
+  - gfortran_linux-64=9.5.0=hdb51d14_10
+  - graphite2=1.3.13=h58526e2_1001
+  - gsl=2.7=he838d99_0
+  - gxx_impl_linux-64=9.5.0=h6c5bc03_16
+  - gxx_linux-64=9.5.0=h43f449f_10
+  - harfbuzz=4.2.0=h40b6f09_0
+  - icu=69.1=h9c3ff4c_0
+  - jpeg=9e=h166bdaf_2
+  - kernel-headers_linux-64=2.6.32=he073ed8_15
+  - keyutils=1.6.1=h166bdaf_0
+  - krb5=1.19.3=h08a2579_0
+  - ld_impl_linux-64=2.36.1=hea4e1c9_2
+  - lerc=4.0.0=h27087fc_0
+  - libblas=3.9.0=16_linux64_openblas
+  - libcblas=3.9.0=16_linux64_openblas
+  - libcurl=7.85.0=h2283fc2_0
+  - libdeflate=1.14=h166bdaf_0
+  - libedit=3.1.20191231=he28a2e2_2
[…]
