### 2022-10-11 NorwegianVeterinaryInstitute/ALPPACA#28 @hkaspersen
Pipeline can now be run with conda,singularity, and docker
