### 2022-10-12 191814f
updated text
Changes based on issue #39 and #40 (state of the field, ie description of overlapping pipelines as of https://github.com/openjournals/joss-reviews/issues/4677#issuecomment-1242767318, now incorporated in the statement of need), and changes to the tracks section based on issue #41.
