+++ paper.md
@@ -20,23 +20,25 @@ bibliography: paper.bib
 ---
 
 # Summary
+
+
 Phylogenetic analysis is often a complex and time-consuming process. We developed A tooL for Prokaryotic Phylogeny And Clustering Analysis (ALPPACA), a pipeline designed for phylogenetic analysis of prokaryotes. The pipeline is implemented in Nextflow and utilize singularity containers, which makes the pipeline portable and reproducible. The pipeline consists of three different tracks designed to cover three different levels of expected dataset diversity, facilitating the choice of performing appropriate analysis for different datasets. The final results are presented in a tidy HTML report for each track. The code, documentation and examples of configuration files to run analyses with are publicly available on Github.
 
 # Statement of need
-Unraveling the evolutionary relationship between organisms is a crucial part of many comparative genomics projects. However, the complexity of preparing, running, and interpreting such analyses is a hindrance for many researchers not specialized in evolutionary biology. Additionally, choosing a combination of compatible software for various analysis scenarios may be difficult. Here we present A tooL for Prokaryotic Phylogeny And Clustering Analysis (ALPPACA), a Nextflow [@DiTommaso:2017] pipeline for phylogenetic analysis of prokaryotic genomes. The pipeline has been developed to make it easier to run phylogenetic analysis and provides reports to simplify result interpretation. Our pipeline is composed of three main tracks, designed to allow analysis of datasets represented by three different genetic diversity levels. These levels of similarity influence what assumptions are used to consider sequences as orthologous when reconstructing the multiple alignment required for phylogenetic inference. For instance, our pipeline can be run on datasets with an expected low- (eg. outbreak situation), medium- (eg. within a single MLST cluster), or high diversity level (eg. across ST or for closely related species). One major advantage of this pipeline is the possibility for the user to select tracks based on their data at hand, which makes it possible to generate analysis results in a rapid fashion.
+Unraveling the evolutionary relationship between organisms is a crucial part of many comparative genomics projects. Phylogenetic analysis often entails running several tools consecutively, and several assumptions are made for the data used in each tool. To add to this complexity, several tools have been developed for each step in such an analysis, and sifting through these tools as a user may be time-consuming and difficult. Additionally, choosing a combination of compatible software for various analysis scenarios may require in-depth knowledge and experience in the field of microbial evolution. To the authors knowledge, currently the only available Nextflow [@DiTommaso:2017] pipeline for a start-to-end phylogenetic analysis is bactmap (https://nf-co.re/bactmap). However, bactmap only allows reference-based phylogenetic analysis by mapping reads to a reference. Here, we present A tooL for Prokaryotic Phylogeny And Clustering Analysis (ALPPACA), a Nextflow pipeline for phylogenetic analysis of prokaryotic genomes. ALPPACA provides a suite of analyses tailored for different scenarios, designed to allow analysis of datasets represented by three different genetic diversity levels, all in one package. These levels of similarity influence what assumptions are used to consider sequences as orthologous when reconstructing the multiple alignment required for phylogenetic inference.
 
 # Pipeline and track descriptions
 ## Pipeline
[…]
