### 2022-10-12 7b021ad
updated conclusion
made changes to the conclusion, as of issue #41
