### 2022-10-12 706f51d
Updated statement of need
added small section, as of issue #39
