### 2022-10-13 c7c1037
added additional references
