+++ paper.bib
@@ -54,8 +54,7 @@
   volume = {35},
   number = {4},
   pages = {316--319},
-  doi = {10.1038/nbt.3820},
-  url = {http://www.nature.com/articles/nbt.3820}
+  doi = {10.1038/nbt.3820}
 }
 @article{Treangen:2014,
   title = {The Harvest Suite for Rapid Core-Genome Alignment and Visualization of Thousands of Intraspecific Microbial Genomes},
@@ -65,8 +64,7 @@
   volume = {15},
   number = {11},
   pages = {524},
-  doi = {10.1186/s13059-014-0524-x},
-  url = {http://genomebiology.biomedcentral.com/articles/10.1186/s13059-014-0524-x}
+  doi = {10.1186/s13059-014-0524-x}
 }
 @article{Croucher:2015,
   title = {Rapid Phylogenetic Analysis of Large Samples of Recombinant Bacterial Whole Genome Sequences Using Gubbins},
@@ -76,6 +74,35 @@
   volume = {43},
   number = {3},
   pages = {e15--e15},
-  doi = {10.1093/nar/gku1196},
-  url = {http://academic.oup.com/nar/article/43/3/e15/2410982/Rapid-phylogenetic-analysis-of-large-samples-of}
+  doi = {10.1093/nar/gku1196}
+}
+@article{Kaspersen:2020,
+  title = {Dissemination of Quinolone-Resistant Escherichia coli in the Norwegian Broiler and Pig Production Chains and Possible Persistence in the Broiler Production Environment},
+  author = {Kaspersen, H. and Sekse, C. and Fiskebeck, E. Z. and Slettemeås, J. S. and Simm, R. and Norström, M. and Urdahl, A. M. and Lagesen, K.},
+  date = {2020-01},
+  journaltitle = {Applied and Environmental Microbiology},
+  volume = {86},
+  number = {7},
+  pages = {e02769-19},
+  doi = {10.1128/AEM.02769-19}
+}
+@article{Franklin-Alming:2021,
+  title = {Exploring Klebsiella pneumoniae in Healthy Poultry Reveals High Genetic Diversity, Good Biofilm-Forming Abilities and Higher Prevalence in Turkeys Than Broilers},
+  author = {Franklin-Alming, F. and Kaspersen, H. and Hetland, M. A. K. and Bakksjø, R. and Nesse, L. L. and Leangapichart, T. and Löhr, I. H. and Telke, A. A. and Sunde, M.},
+  date = {2021-09},
+  journaltitle = {Frontiers in Microbiology},
+  volume = {12},
+  number = {725414},
+  pages = {11},
+  doi = {10.3389/fmicb.2021.725414}
+}
+@article{Smistad:2022,
+  title = {Streptococcus Dysgalactiae Ssp. Dysgalactiae in Norwegian Bovine Dairy Herds: Risk Factors, Sources, and Genomic Diversity},
+  author = {Smistad, M. and Kaspersen, H. and Franklin-Alming, F.V. and Wolff, C. and Sølverød, L. and Porcellato, D. and Trettenes, E. and Jørgensen, H.J.},
+  date = {2022-04},
+  journaltitle = {Journal of Dairy Science},
+  volume = {105},
+  number = {4},
+  pages = {3574--3587},
+  doi = {10.3168/jds.2021-21471}
 }
