+++ paper.bib
@@ -106,3 +106,13 @@
   pages = {3574--3587},
   doi = {10.3168/jds.2021-21471}
 }
+@article{Kravik:2022,
+  title = {High Sequence Similarity between Avian Pathogenic E. coli Isolates from Individual Birds and within Broiler Chicken Flocks during Colibacillosis Outbreaks},
+  author = {Kravik, I. H. and Kaspersen, H. and Sjurseth, S. K. and Jonsson, M. and David, B. and Aspholm, M. and Sekse, C.},
+  date = {2022-04},
+  journaltitle = {Veterinary Microbiology},
+  volume = {267},
+  number = {109378},
+  pages = {109378},
+  doi = {10.1016/j.vetmic.2022.109378}
+}
