### 2022-10-13 4154dd7
Additions and review
Re-wrote the summary as of issue #38, and made changes to statement of need to reflect the new summary as in #39. Added to the conclusion as of issue #41.
