+++ paper.md
@@ -4,7 +4,7 @@ tags:
   - Nextflow
   - Phylogeny
   - Prokaryote
-  - R
+  - Clustering
 authors:
   - name: Håkon Kaspersen
     orcid: 0000-0002-9559-1303
@@ -20,41 +20,32 @@ bibliography: paper.bib
 ---
 
 # Summary
-
-
-Phylogenetic analysis is often a complex and time-consuming process. We developed A tooL for Prokaryotic Phylogeny And Clustering Analysis (ALPPACA), a pipeline designed for phylogenetic analysis of prokaryotes. The pipeline is implemented in Nextflow and utilize singularity containers, which makes the pipeline portable and reproducible. The pipeline consists of three different tracks designed to cover three different levels of expected dataset diversity, facilitating the choice of performing appropriate analysis for different datasets. The final results are presented in a tidy HTML report for each track. The code, documentation and examples of configuration files to run analyses with are publicly available on Github.
+Phylogenetic analysis is an important method that has been used to unravel outbreaks and track the origin of pathogens worldwide. Phylogenetic analysis has also become commonplace in several research projects and clinical investigations, where time is often of the essence. This kind of analysis often entails running several tools consecutively, and several assumptions are made for the data used in each tool. To add to this complexity, several tools have been developed for each step in such an analysis, and sifting through these tools as a user may be time-consuming and difficult. Additionally, choosing a combination of compatible software for various analysis scenarios may require in-depth knowledge and experience in the field of microbial evolution, which often prevents non-specialists from utilizing such analyses. Here we present a solution that will help alleviate these problems in hopes of making it easier and faster to run reproducible phylogenetic analyses.
 
 # Statement of need
-Unraveling the evolutionary relationship between organisms is a crucial part of many comparative genomics projects. Phylogenetic analysis often entails running several tools consecutively, and several assumptions are made for the data used in each tool. To add to this complexity, several tools have been developed for each step in such an analysis, and sifting through these tools as a user may be time-consuming and difficult. Additionally, choosing a combination of compatible software for various analysis scenarios may require in-depth knowledge and experience in the field of microbial evolution. To the authors knowledge, currently the only available Nextflow [@DiTommaso:2017] pipeline for a start-to-end phylogenetic analysis is bactmap (https://nf-co.re/bactmap). However, bactmap only allows reference-based phylogenetic analysis by mapping reads to a reference. Here, we present A tooL for Prokaryotic Phylogeny And Clustering Analysis (ALPPACA), that allows both denovo and reference-based phylogenetic  
-analysis of prokaryotic genomes. The Nextflow pipeline ALPPACA provides a suite of analyses tailored for different scenarios, designed to allow analysis of datasets represented by three different genetic diversity levels, all in one package. These levels of similarity influence what assumptions are used to consider sequences as orthologous when reconstructing the multiple alignment required for phylogenetic inference. The tracks differs in their approach to recover the multiple genome (MGA) or
-multi-gene (MGenA) alignments that are used for phylogenetic inference. When genomes similarity is expected to be low, the core gene track annotates genomes, which allows identifying core orthologous genes and produces a concatenated MGenA. While the mapping track is suitable for more similar genomes. The reference used for mapping may either be one of the genomes under investigation, or a closely related reference of choice, and serves as a common referential for SNPs detection which allow reconstructing a MGA of all the loci that are common with the reference. Lastly, the core genome
-track, suitable with only very closely related genomes, allows reconstruction of a MGA by aligning genome sequences. It first find unique sequences that are common all
[…]
