### 2022-10-13 #4677 @hkaspersen
@csoneson how do I format bacterial names correctly in the `.bib` file, they are not in italics in the preview pdf. Can I use `\textit` bibtex formatting?
