### 2022-10-14 db65d84
Suggestion improvement article joss
Suggested changes and questions
