+++ paper.md
@@ -32,7 +32,7 @@ The ALPPACA pipeline is written in Nextflow [@DiTommaso:2017], and the code and 
 ## Tracks
 The pipeline consists of three separate tracks depending on the objectives and data available to the user (\autoref{fig:figure1}). Each track differ in the way they detect homologous regions to construct the multiple alignment needed for phylogenetic inference.
 
-First, the core gene track is designed to be used for datasets that are expected to have a relatively high level of genetic diversity. This track is useful if you have a dataset with different but closely related species, or different STs of the same species. The track provides the means to generate a relatively high resolution phylogeny on diverse datasets. Here, the genomes are annotated with Prokka [@Seeman:2014], the pangenome is inferred using Panaroo [@Tonkin-Hill:2020], and a phylogenetic reconstruction is generated with IQTree [@Nguyen:2015] from a core gene alignment.
+First, the core gene track is designed to be used for datasets that are expected to have a relatively high level of genetic diversity. This track is useful if you have a dataset with different but closely related species, or different STs of the same species. The track provides the means to generate a relatively high resolution phylogeny on diverse datasets. Here, the genomes are annotated with Prokka [@Seeman:2014], the pangenome is inferred using Panaroo [@Tonkin-Hill:2020], a core gene alignment is produced and used for phylogenetic reconstruction with IQTree [@Nguyen:2015].
 
 The mapping track is designed for datasets that are expected to have a medium to low diversity level. This track maps reads to a reference using Snippy (https://github.com/tseemann/snippy). Mapping reads circumvent the need to generate assemblies, which is a time-consuming process. Time is usually of the essence in an outbreak situation, and this track allows for rapid analysis of outbreak data. Additionally, this track ensures that only genetic material from vertical descent is included in the analysis, as recombinant areas are detected by Gubbins [@Croucher:2015] and masked with Maskrc-svg (https://github.com/kwongj/maskrc-svg) before the phylogeny is inferred with IQTree. 
 
