### 2022-10-14 0b814eb
maybe language improvement
PS: We might want had some experience in the doc for what to expect in genome coverage eg 95 % and above when extremely similar osv as indication for users
Based on our experience
