+++ paper.md
@@ -36,7 +36,7 @@ First, the core gene track is designed to be used for datasets that are expected
 
 The mapping track is designed for datasets that are expected to have a medium to low diversity level. This track maps reads to a reference using Snippy (https://github.com/tseemann/snippy). Mapping reads circumvent the need to generate assemblies, which is a time-consuming process. Time is usually of the essence in an outbreak situation, and this track allows for rapid analysis of outbreak data. Additionally, this track ensures that only genetic material from vertical descent is included in the analysis, as recombinant areas are detected by Gubbins [@Croucher:2015] and masked with Maskrc-svg (https://github.com/kwongj/maskrc-svg) before the phylogeny is inferred with IQTree. 
 
-Lastly, the core genome track is designed to be used for datasets that are expected to have a low level of genetic diversity, e.g. within the same ST. This track is useful after identifying clusters of interest using the core gene track above, to increase the resolution of the phylogenetic reconstruction on a subset of the dataset. This track outputs the percent length of each genome included in the alignment, reported as average genome coverage by ParSNP [@Treangen:2014]. This is an important parameter to consider when interpreting the results, as it tells the user how much of each genome the phylogenetic inference is based on. Similar to the mapping track, only data from vertical descent is included in the analysis.
+Lastly, the core genome track is designed to be used for datasets that are expected to have a low level of genetic diversity, e.g. within the same ST. This track is useful after identifying clusters of interest using the core gene track above, to increase the resolution of the phylogenetic analysis on a subset of the dataset. This track outputs the percent length of each genome included in the alignment, reported as average genome coverage by ParSNP [@Treangen:2014]. This is an important parameter to consider when interpreting the results, as it tells the user how much of each genome the phylogenetic inference is based on. Similar to the mapping track, only data from vertical descent is included in the analysis.
 
 All three tracks also generate SNP distance statistics with snp-dists (https://github.com/tseemann/snp-dists). The SNP distances are very useful when defining clusters, or if defining outbreak clades based on a SNP distance cutoff. The user also has options to deduplicate the alignment with Seqkit [@Shen:2016], and filter out constant sites with snp-sites [@Page:2016], which will reduce runtime of the pipeline.
 
