+++ paper.md
@@ -43,7 +43,7 @@ All three tracks also generate SNP distance statistics with snp-dists (https://g
 ![Overview of the three tracks in ALPPACA.\label{fig:figure1}](pipeline.png)
 
 # Conclusion
-The ALPPACA pipeline provides a suite of phylogenetic analyses for different scenarios, all in one package. This enables a variety of uses without having to download several tools and programs, and the Nextflow framework allows for user-friendly and reproducible use of the pipeline. Additional tracks may be added to ALPPACA in the future, such as clusering based on core/whole genome multi locus sequence typing, or additions to existing tracks, such as recombination detection in the core gene analysis. Clustering analysis using FastANI (https://github.com/ParBLiSS/FastANI) will be added as a separate track, to aid users in selecting the correct track for their data at hand.
+The ALPPACA pipeline provides a suite of phylogenetic analyses for different scenarios, all in one package. This enables a variety of uses without having to download several tools and programs, and the Nextflow framework allows for user-friendly and reproducible use of the pipeline. Additional tracks may be added to ALPPACA in the future, such as clusering based on core/whole genome multi locus sequence typing, or additions to existing tracks, such as recombination detection in the core gene analysis. Clustering analysis using FastANI (https://github.com/ParBLiSS/FastANI) will be added as a separate track, to aid users in selecting the correct track by evaluating genetic diversity in their data.
 
 # Acknowledgements
 The projects QREC-MaP (Research Funding for Agriculture and the Food Industry, Norwegian Research Council, project number 255383), KLEB-GAP (Trond Mohn Foundation, project number TMS2019TMT03), Yersiniosis at Sea (Norwegian Seafood Research Fund grant, project number 901505), and "Increasing sustainability of Norwegian food production by tackling streptococcal infections in modern livestock systems" (Norwegian Agricultural Agreement Research Fund, Norwegian Research council) are acknowledged for providing the research platform for this work. The computations were performed on resources provided by UNINETT Sigma2 - the National Infrastructure for High Performance Computing and Data Storage in Norway.
