### 2022-10-14 NorwegianVeterinaryInstitute/ALPPACA#35 @hkaspersen
Comment on how to implement testing:
https://github.com/openjournals/joss-reviews/issues/4677#issuecomment-1272026591
In NF-CORE pipelines, testing is implemented using a `test.config` file that defines the input and parameters of the test.
However, as far as I can see, I don't understand fully how Nextflow handles a link to files online, and converts it to a channel.
Example from the [viralrecon](https://github.com/nf-core/viralrecon/blob/master/conf/test.config) pipeline:
```
// Input data to test amplicon analysis
    input              = 'https://raw.githubusercontent.com/nf-core/test-datasets/viralrecon/samplesheet/samplesheet_test_illumina_amplicon.csv'
    platform           = 'illumina'
    protocol           = 'amplicon'
```
here they use a samplesheet, and they have some nextflow processes that handle this, but I cannot see where it is converted to a channel for the processes after.
@evezeyl if you can look further into this i'd be grateful! Added @karinlag too in case you have any additional insight here.
