### 2022-10-14 NorwegianVeterinaryInstitute/ALPPACA#35 @evezeyl
As far as I can see in their pipeline they create the Chanel from the params.input which is the paths for a CSV file

In the test config you provide the path ... Just that it's online.
