### 2022-10-17 #4677 @csoneson
Hi @hkaspersen - yes, I believe you should be able to use `\textit` in the titles in your `.bib` file.
