### 2022-10-18 NorwegianVeterinaryInstitute/ALPPACA#35 @hkaspersen
Not yet managed to get the csv-file split into fasta channels. Asked on Nextflow Discussion:
https://github.com/nextflow-io/nextflow/discussions/3294
