### 2022-10-21 d7e02a6
added channel factory for input csv file, to allow testing with profile, as of issue #35
