### 2022-10-21 NorwegianVeterinaryInstitute/ALPPACA#35 @hkaspersen
The solution was to use the following:
```
fasta_channel = Channel
    .fromPath(params.input)
    .splitCsv(header:true, sep:",")
    .map { file(it.path) }
    .collect()
```
This will pull the fasta files and create one channel for all of them together
