### 2022-10-21 NorwegianVeterinaryInstitute/ALPPACA#35 @hkaspersen
Test (core genome) now functional
