+++ paper.md
@@ -20,7 +20,7 @@ bibliography: paper.bib
 ---
 
 # Summary
-Phylogenetic analysis is an important method that has been used to unravel outbreaks and <!-- what do you mean? --> track the origin of pathogens worldwide. Phylogenetic analysis has also become commonplace in several research projects and clinical investigations, where time is often of the essence. This kind of analysis often entails running several tools consecutively, and several assumptions are made for the data used in each tool. To add to this complexity, several tools have been developed for each step in such an analysis, and sifting through these tools as a user may be time-consuming. Additionally, choosing a combination of compatible software for various analysis scenarios may require in-depth knowledge and experience in the field of microbial evolution, which often prevents non-specialists from utilizing such analyses. Here we present a solution that will help alleviate these problems in hopes of making it easier and faster to run reproducible phylogenetic analyses.
+Phylogenetic analysis is an important method that has been used to unravel outbreaks and track the origins of pathogens worldwide. Phylogenetic analysis has also become commonplace in several research projects and clinical investigations, where time is often of the essence. This kind of analysis often entails running several tools consecutively, and several assumptions are made for the data used in each tool. To add to this complexity, several tools have been developed for each step in such an analysis, and sifting through these tools as a user may be time-consuming. Additionally, choosing a combination of compatible software for various analysis scenarios may require in-depth knowledge and experience in the field of microbial evolution, which often prevents non-specialists from utilizing such analyses. Here we present a solution that will help alleviate these problems in hopes of making it easier and faster to run reproducible phylogenetic analyses.
 
 # Statement of need
 A tooL for Prokaryotic Phylogeny And Clustering Analysis (ALPPACA) is a pipeline that allows both de-novo and reference-based phylogenetic analysis of prokaryotic genomes. The pipeline provides a suite of analyses tailored for different scenarios, designed to allow analysis of datasets represented by three different genetic diversity levels, all in one package. These levels of similarity influence what assumptions are used to consider sequences as orthologous when reconstructing the multiple alignment required for phylogenetic inference. By selecting an appropriate track for the data at hand, the user can be confident that these assumptions are taken care of within the framework of ALPPACA. The ability to run different datasets through different tracks within the same framework makes ALPPACA unique compared to other phylogeny pipelines such as Bactmap (https://nf-co.re/bactmap) and SNVPhyl (https://github.com/DHQP/SNVPhyl_Nextflow), where only mapping-based phylogeny is possible. Several research projects have generated the framework of ALPPACA, and and peer-reviewed papers have been published using this framework for phylogenetic analysis [@Kaspersen:2020;@Franklin-Alming:2021;@Smistad:2022;@Kravik:2022].
@@ -32,7 +32,7 @@ The ALPPACA pipeline is written in Nextflow [@DiTommaso:2017], and the code and 
 ## Tracks
 The pipeline consists of three separate tracks depending on the objectives and data available to the user (\autoref{fig:figure1}). Each track differ in the way they detect homologous regions to construct the multiple alignment needed for phylogenetic inference.
 
-First, the core gene track is designed to be used for datasets that are expected to have a relatively high level of genetic diversity. This track is useful if you have a dataset with different but closely related species, or different STs of the same species. The track provides the means to generate a relatively high resolution phylogeny on diverse datasets. Here, the genomes are annotated with Prokka [@Seeman:2014], the pangenome is inferred using Panaroo [@Tonkin-Hill:2020], a core gene alignment is produced and used for phylogenetic reconstruction with IQTree [@Nguyen:2015].
[…]
