### 2022-10-21 8bd1a63
fixed italic bacterial names
