+++ paper.bib
@@ -77,7 +77,7 @@
   doi = {10.1093/nar/gku1196}
 }
 @article{Kaspersen:2020,
-  title = {Dissemination of Quinolone-Resistant Escherichia coli in the Norwegian Broiler and Pig Production Chains and Possible Persistence in the Broiler Production Environment},
+  title = {Dissemination of Quinolone-Resistant \textit{Escherichia coli} in the Norwegian Broiler and Pig Production Chains and Possible Persistence in the Broiler Production Environment},
   author = {Kaspersen, H. and Sekse, C. and Fiskebeck, E. Z. and Slettemeås, J. S. and Simm, R. and Norström, M. and Urdahl, A. M. and Lagesen, K.},
   date = {2020-01},
   journaltitle = {Applied and Environmental Microbiology},
@@ -87,7 +87,7 @@
   doi = {10.1128/AEM.02769-19}
 }
 @article{Franklin-Alming:2021,
-  title = {Exploring Klebsiella pneumoniae in Healthy Poultry Reveals High Genetic Diversity, Good Biofilm-Forming Abilities and Higher Prevalence in Turkeys Than Broilers},
+  title = {Exploring \textit{Klebsiella pneumoniae} in Healthy Poultry Reveals High Genetic Diversity, Good Biofilm-Forming Abilities and Higher Prevalence in Turkeys Than Broilers},
   author = {Franklin-Alming, F. and Kaspersen, H. and Hetland, M. A. K. and Bakksjø, R. and Nesse, L. L. and Leangapichart, T. and Löhr, I. H. and Telke, A. A. and Sunde, M.},
   date = {2021-09},
   journaltitle = {Frontiers in Microbiology},
@@ -97,7 +97,7 @@
   doi = {10.3389/fmicb.2021.725414}
 }
 @article{Smistad:2022,
-  title = {Streptococcus Dysgalactiae Ssp. Dysgalactiae in Norwegian Bovine Dairy Herds: Risk Factors, Sources, and Genomic Diversity},
+  title = {\textit{Streptococcus Dysgalactiae} Ssp. \textit{dysgalactiae} in Norwegian Bovine Dairy Herds: Risk Factors, Sources, and Genomic Diversity},
   author = {Smistad, M. and Kaspersen, H. and Franklin-Alming, F.V. and Wolff, C. and Sølverød, L. and Porcellato, D. and Trettenes, E. and Jørgensen, H.J.},
   date = {2022-04},
   journaltitle = {Journal of Dairy Science},
@@ -107,7 +107,7 @@
   doi = {10.3168/jds.2021-21471}
 }
 @article{Kravik:2022,
-  title = {High Sequence Similarity between Avian Pathogenic {{E}}. coli Isolates from Individual Birds and within Broiler Chicken Flocks during Colibacillosis Outbreaks},
+  title = {High Sequence Similarity between Avian Pathogenic \textit{E. coli} Isolates from Individual Birds and within Broiler Chicken Flocks during Colibacillosis Outbreaks},
   author = {Kravik, I. H. and Kaspersen, H. and Sjurseth, S. K. and Jonsson, M. and David, B. and Aspholm, M. and Sekse, C.},
   date = {2022-04},
   journaltitle = {Veterinary Microbiology},
