+++ paper.bib
@@ -97,7 +97,7 @@
   doi = {10.3389/fmicb.2021.725414}
 }
 @article{Smistad:2022,
-  title = {\textit{Streptococcus dysgalactiae} Ssp. \textit{dysgalactiae} in Norwegian Bovine Dairy Herds: Risk Factors, Sources, and Genomic Diversity},
+  title = {\textit{Streptococcus dysgalactiae} Ssp. \textit{{{d}}ysgalactiae} in Norwegian Bovine Dairy Herds: Risk Factors, Sources, and Genomic Diversity},
   author = {Smistad, M. and Kaspersen, H. and Franklin-Alming, F.V. and Wolff, C. and Sølverød, L. and Porcellato, D. and Trettenes, E. and Jørgensen, H.J.},
   date = {2022-04},
   journaltitle = {Journal of Dairy Science},
