+++ bin/core_genome_report.Rmd
@@ -19,8 +19,9 @@ params:
     value: x
 ---
 
-# Version information
-ALPPACA version 1.0.0.
+# Run information
+
+Date generated: `r Sys.Date()`
 
 ## R Packages
 These are the R packages used to generate this report.
@@ -72,7 +73,7 @@ tool_data <- data.frame(
            "Snp-sites",
            "IQTree"),
   Version = c("1.6.1",
-              "3.1.3",
+              "3.2.0",
               "0.5",
               "0.8.2",
               "2.5.1",
