+++ bin/mapping_report.Rmd
@@ -22,7 +22,7 @@ params:
 
 Date generated: `r Sys.Date()`
 
-## R Packages used
+## R Packages
 These are the R packages used to generate this report.
 ```{r, message=FALSE, warning=FALSE}
 library(stringr)
@@ -72,7 +72,7 @@ tool_data <- data.frame(
            "Snp-sites",
            "IQTree"),
   Version = c("4.6.0",
-              "3.2",
+              "3.2.0",
               "0.5",
               "0.8.2",
               "2.5.1",
