### 2022-11-01 a511ea7
added script for generating input file
