### 2022-11-01 #4677 @hkaspersen
Dear @csoneson @rcannood @mberacochea @hseabolt.

Thank you so much for your valuable feedback on ALPPACA!

I have now worked on the changes mentioned by you all, and created a new release of ALPPACA (v.2.0.0). The changes are reflected in the issues mentioned above, and in the release description. @rcannood I will keep your issue open until the review is completed.

The main changes include the following:
- Wrapper scripts have been removed. Pipeline is now executed directly with `nextflow run main.nf` or directly from github using `nextflow run NorwegianVeterinaryInstitute/ALPPACA`
- Input is now a `csv` file with sample ids and full paths to all files used. This file can be created with the script supplied at `bin/generate_input.R`
- Singularity, docker, and conda is supported as image handlers (specified with `-profile`)
- Test functionality added using `-profile test`
- Example configs are supplied for local and HPC systems that users can modify for their system
- Default parameter values are specified in `nextflow.config`
- Updates to the Wiki to reflect the changes in usage above. Link to wiki have now been specified on the main page of the pipeline
- `yml` files added which users can use to generate a `nextflow` conda environment for running the pipeline

The paper has also been updated to reflect your suggestions above.

@csoneson let me know if I have to supply more information here!
Looking forward to hear from you again!
