### 2022-11-10 NorwegianVeterinaryInstitute/ALPPACA#27 @rcannood
Hey @hkaspersen! Congratulations on addressing all of my comments in such a brief period of time! :)

I believe the documentation, the software repository and the paper all look _so_ much better now.

I have a minor comment on the structure of the manuscript, but I'll pick this up in a different issue.
