### 2022-11-10 NorwegianVeterinaryInstitute/ALPPACA#27 @rcannood
Actually, let me reopen this issue ;)

In #35 I see you added a way for testing the pipeline by defining a test profile.

I'm happy there is a test for the pipeline, however the test is not _automated_ just yet.

To fully automate your test, you can browse some of the nf core pipelines and see how they implemented CI. For example, the isoseq pipeline has one: https://github.com/nf-core/isoseq/blob/master/.github/workflows/ci.yml.
