### 2022-11-11 NorwegianVeterinaryInstitute/ALPPACA#27 @hkaspersen
Dear @rcannood, thank you for your feedback!
I am not completely sure what you mean with automated in this context. Do you mean that the user should not need to actually run the command to test the pipeline? Or is this a test run to check if the pipeline runs after I am updating it?

If you feel it is necessary for me to implement this I am happy to oblige, but I need to remind you that this is not (yet) an NF-Core pipeline, so I just want you to be absolutely sure it is needed! Could you perhaps elaborate on the usefulness of such a functionality?

Thanks in advance!
