### 2022-11-11 NorwegianVeterinaryInstitute/ALPPACA#27 @rcannood
I'm referring to the NF core pipeline as an example. An explanation on automated testing can be found in the JOSS documentation here: https://joss.readthedocs.io/en/latest/review_criteria.html#tests . The NF core pipeline I linked to is an example of an automated testing suite using GitHub actions.
