### 2022-11-11 NorwegianVeterinaryInstitute/ALPPACA#27 @hkaspersen
@rcannood I was unaware that this was a criteria from JOSS! Thank you for informing me. I will start implementing this as soon as possible.
