### 2022-11-13 ad43fa1
added automated test run
