+++ .github/workflows/ci.yml
@@ -1,4 +1,4 @@
-name: nf-core CI
+name: ALPPACA CI
 # This workflow runs the pipeline with the minimal test dataset to check that it completes without any syntax errors
 on:
   push:
