### 2022-11-13 d7ef892
tried another strategy
