### 2022-11-13 NorwegianVeterinaryInstitute/ALPPACA#35 @hkaspersen
Made test automatic when pushing to the dev branch. Check test before merging with master!
