### 2022-11-13 NorwegianVeterinaryInstitute/ALPPACA#27 @rcannood
GJ! 😀
