### 2022-11-13 NorwegianVeterinaryInstitute/ALPPACA#45 @hkaspersen
Thank you for your feedback @rcannood! Yes, I agree with your interpretation when reading it again. I will restructure the manuscript and add a small summary section at the start. I think I got a bit confused about the summary section due to the [example manuscript](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography) that JOSS provide.
