### 2022-11-13 NorwegianVeterinaryInstitute/ALPPACA#45 @hkaspersen
@rcannood reading it yet again made me think - perhaps the "Prokaryotic Phylogeny and Clustering Analysis" section can actually work as the summary? Moving it up makes sense, as it is supposed to provide a "high level description of the functionality of the software", which it does. I am very close to the word limit, so I'm a bit reluctant to add a while new section.
