### 2022-11-13 #4677 @hkaspersen
@csoneson how strict is JOSS on the structure of the paper? I'm having some issues with the sections and the max word limit - and also tryint to avoid repeating myself too much. As far as I have understood, the Summary section is not equivalent to an Abstract. In my manuscript though it sort of is intertwined with the statement of need. Is it possible to merge these two into one summary? Or is it mandatory to include both headers?
