### 2022-11-13 NorwegianVeterinaryInstitute/ALPPACA#45 @hkaspersen
@evezeyl any thought on this?
