+++ bin/core_genome_report.Rmd
@@ -77,7 +77,7 @@ tool_data <- data.frame(
               "0.5",
               "0.8.2",
               "2.5.1",
-              "2.1.4")
+              "2.2.0.3")
 )
 
 tool_data %>%
