### 2022-11-14 c08e158
structure update as of issue #45
