### 2022-11-14 NorwegianVeterinaryInstitute/ALPPACA#45 @hkaspersen
@rcannood See if you like the new structure better! :)
