### 2022-11-15 NorwegianVeterinaryInstitute/ALPPACA#45 @evezeyl
@hkaspersen I had a look, this was a great suggestion to reorganize as such. Its all good for me :)
