### 2022-11-15 #4677 @hkaspersen
Just to update the other reviewers:
I have changed the structure of the paper to reflect the suggestions from @rcannood, and also made the pipeline test run automatically on push to the dev branch.
