### 2022-11-15 #4677 @csoneson
> @csoneson how strict is JOSS on the structure of the paper? I'm having some issues with the sections and the max word limit - and also trying to avoid repeating myself too much. As far as I have understood, the Summary section is not equivalent to an Abstract. In my manuscript though it sort of is intertwined with the statement of need. Is it possible to merge these two into one summary? Or is it mandatory to include both headers?

@hkaspersen - I'd say that the most important is that both of these sections, with different purposes, are covered in the paper (from the docs):
- A _summary_ describing the high-level functionality and purpose of the software for a diverse, non-specialist audience.
- A _Statement of need_ section that clearly illustrates the research purpose of the software and places it in the context of related work.

Often it may be easier to provide them as separate sections. However, if you think that it can be well structured under a joint header ('Summary and Statement of Need'), that would cover both these aspects, that is also acceptable.
