### 2022-11-15 88f409f
added core_gene to github ci test, and renamed core_genome test
