+++ .github/workflows/ci.yml
@@ -36,6 +36,9 @@ jobs:
         run: |
           wget -qO- get.nextflow.io | bash
           sudo mv nextflow /usr/local/bin/
-      - name: Run pipeline with test data
+      - name: Run core genome track test
         run: |
-          nextflow run ${GITHUB_WORKSPACE} -profile test,docker --out_dir ./results
+          nextflow run ${GITHUB_WORKSPACE} -profile test_core_genome,docker --out_dir ./results_core_genome
+      - name: Run core gene track test
+        run: |
+	  nextflor run ${GITHUB_WORKSPACE} -profile test_core_gene,docker --out_dir ./results_core_gene
