+++ .github/workflows/ci.yml
@@ -41,4 +41,4 @@ jobs:
           nextflow run ${GITHUB_WORKSPACE} -profile test_core_genome,docker --out_dir ./results_core_genome
       - name: Run core gene track test
         run: |
-	  nextflor run ${GITHUB_WORKSPACE} -profile test_core_gene,docker --out_dir ./results_core_gene
+	  nextflow run ${GITHUB_WORKSPACE} -profile test_core_gene,docker --out_dir ./results_core_gene
