+++ .github/workflows/ci.yml
@@ -33,9 +33,11 @@ jobs:
         run: |
           wget -qO- get.nextflow.io | bash
           sudo mv nextflow /usr/local/bin/
+
       - name: Run core genome track test
         run: |
           nextflow run ${GITHUB_WORKSPACE} -profile test_core_genome,docker --out_dir ./results_core_genome
+
       - name: Run core gene track test
         run: |
 	  nextflow run ${GITHUB_WORKSPACE} -profile test_core_gene,docker --out_dir ./results_core_gene
