### 2022-11-16 8f92b0f
added mapping test run
