+++ .github/workflows/ci.yml
@@ -41,3 +41,8 @@ jobs:
       - name: Run core gene track test
         run: |
           nextflow run ${GITHUB_WORKSPACE} -profile test_core_gene,docker --out_dir ./results
+
+     - name: Run mapping track test
+       run: |
+          nextflow run ${GITHUB_WORKSPACE} -profile test_mapping,docker --out_dir ./results
+
