### 2022-11-17 3ee5a35
made qc process optional in the core gene track
