### 2022-11-17 #4677 @hkaspersen
@rcannood thank you so much for your kind words! Your comments were indeed helpful, and I learned a lot!
Just to inform: I bumped the version of the pipeline to 2.0.1. This was due to the addition of the core gene and mapping track tests. Now all three tracks are run when i push to `dev`
