### 2022-11-22 185c7fb
Minor fixes to ms and bib files
