+++ paper.bib
@@ -19,7 +19,7 @@
   doi = {10.1093/bioinformatics/btu153}
 }
 @article{Tonkin-Hill:2020,
-  title = {Producing Polished Prokaryotic Pangenomes with the Panaroo Pipeline},
+  title = {Producing Polished Prokaryotic Pangenomes with the {P}anaroo Pipeline},
   author = {Tonkin-Hill, Gerry and MacAlasdair, Neil and Ruis, Christopher and Weimann, Aaron and Horesh, Gal and Lees, John A. and Gladstone, Rebecca A. and Lo, Stephanie and Beaudoin, Christopher and Floto, R. Andres and Frost, Simon D.W. and Corander, Jukka and Bentley, Stephen D. and Parkhill, Julian},
   date = {2020-12},
   journaltitle = {Genome Biology},
@@ -29,7 +29,7 @@
   doi = {10.1186/s13059-020-02090-4}
 }
 @article{Shen:2016,
-  title = {SeqKit: A Cross-Platform and Ultrafast Toolkit for FASTA/Q File Manipulation},
+  title = {SeqKit: A Cross-Platform and Ultrafast Toolkit for {FASTA/Q} File Manipulation},
   author = {Shen, Wei and Le, Shuai and Li, Yan and Hu, Fuquan},
   date = {2016-10},
   journaltitle = {PLOS ONE},
@@ -57,7 +57,7 @@
   doi = {10.1038/nbt.3820}
 }
 @article{Treangen:2014,
-  title = {The Harvest Suite for Rapid Core-Genome Alignment and Visualization of Thousands of Intraspecific Microbial Genomes},
+  title = {The {H}arvest Suite for Rapid Core-Genome Alignment and Visualization of Thousands of Intraspecific Microbial Genomes},
   author = {Treangen, Todd J. and Ondov, Brian D. and Koren, Sergey and Phillippy, Adam M.},
   date = {2014-11},
   journaltitle = {Genome Biology},
@@ -67,7 +67,7 @@
   doi = {10.1186/s13059-014-0524-x}
 }
 @article{Croucher:2015,
-  title = {Rapid Phylogenetic Analysis of Large Samples of Recombinant Bacterial Whole Genome Sequences Using Gubbins},
+  title = {Rapid Phylogenetic Analysis of Large Samples of Recombinant Bacterial Whole Genome Sequences Using {G}ubbins},
   author = {Croucher, Nicholas J. and Page, Andrew J. and Connor, Thomas R. and Delaney, Aidan J. and Keane, Jacqueline A. and Bentley, Stephen D. and Parkhill, Julian and Harris, Simon R.},
   date = {2015-02},
   journaltitle = {Nucleic Acids Research},
@@ -77,7 +77,7 @@
   doi = {10.1093/nar/gku1196}
 }
 @article{Kaspersen:2020,
-  title = {Dissemination of Quinolone-Resistant \textit{{{E}}scherichia coli} in the Norwegian Broiler and Pig Production Chains and Possible Persistence in the Broiler Production Environment},
+  title = {Dissemination of Quinolone-Resistant \textit{{{E}}scherichia coli} in the {N}orwegian Broiler and Pig Production Chains and Possible Persistence in the Broiler Production Environment},
   author = {Kaspersen, H. and Sekse, C. and Fiskebeck, E. Z. and Slettemeås, J. S. and Simm, R. and Norström, M. and Urdahl, A. M. and Lagesen, K.},
   date = {2020-01},
   journaltitle = {Applied and Environmental Microbiology},
@@ -97,7 +97,7 @@
   doi = {10.3389/fmicb.2021.725414}
 }
 @article{Smistad:2022,
[…]
