### 2022-11-22 #4677 @csoneson
👋🏻 @hkaspersen - I have gone through the submission and made a [PR](https://github.com/NorwegianVeterinaryInstitute/ALPPACA/pull/62) with a couple of small fixes. Please check and merge if you agree. Could you also spell out 'ST' the first time you use it in the manuscript? When done, you can generate a new proof as before. The next steps then are as follows:

- [ ] Make a tagged release of the software, and list the version tag of the archived version here.
- [ ] Check that the archival deposit (e.g., in Zenodo) has the correct metadata. In particular, the title and author list should be identical to those of the paper. You can also add the authors' ORCID.
