### 2022-11-22 #4677 @hkaspersen
@csoneson thank you!
I have now merged the PR and updated the Zenodo DOI as mentioned.
Zenodo DOI: 10.5281/zenodo.7331351
Tagged release on GH: v2.0.1
