### 2022-11-22 #4677 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
