### 2022-11-29 #4677 @Kevin-Mattheus-Moerman
@hkaspersen Dear authors, I am the AEiC of this track and here to help process the acceptance of this work in JOSS. I have reviewed the paper, the repository, this review, and the archive on ZENODO and feel everything is in order. I have some comments below but we can proceed to accept this submission now.

- [ ] On your ZENODO archive https://zenodo.org/record/7331351 I recommend that you edit the license listed to match that of your software (the BSD-3-Clause license), the ZENODO archive now lists "Other (open)".
