### 2022-11-29 #4677 @Kevin-Mattheus-Moerman
@hkaspersen congratulations to you and your co-author on your publication in JOSS!! :partying_face:

@csoneson thanks for editing this submission!!

And a special thanks to the reviewers @mberacochea, @hseabolt, and @rcannood!!!! JOSS is not possible without the valuable volunteer efforts by people like yourself, thanks a lot!!!
